Extract every alias and extension method to a common file
